requirements.txt
discord.py>=2.3.2
streamlit>=1.36.0

